app/src/main/python/requirements.txt
familiar-agent[llm,mesh]
